First stab at fixing Ruby 2.7 deprecation warnings.
Ruby 2.7 issues the following warning:
  warning: Using the last argument as keyword parameters is deprecated; maybe ** should be added to the call

The behaviour change is due to plans for Ruby 3 to remove support for
automatic treatment of hashes as keyword arguments. For example, this
code was previously fine but will result in an error in Ruby 3:

```
def double(x:)
  x * 2
end

options = { x: 5 }
double(options)
```

In Ruby 3, an explicit double splat will be required
(`double(**options)`).